Preserve test suite state between runs
Previously, successive invocations would result in separate containers
being made. Record important information--such as the image ID
corresponding to a Dockerfile or the container ID of the last spawned
container--so we don't need to rebuild them each time.

Also included here is logic to return appropriate exit codes based on
whether the tests passed or failed.