Buffer console messages
When executing a command within a Docker container, not every line of
log output ends with a newline. This would result in extra blank lines
in some cases since Chef runs purposely don't output the newline until
the resource has finished converging.

For our purposes, we want to improve the output since we don't need the
absolute-latest output (the log files will still receive that). Buffer
output until a newline character is received so that we can flush with
one fell swoop.