Improve description of suites about to be run
It's nice to get an idea of which suites are about to run at the start
of the test. This is useful in case you forgot to specify a pattern.

We also treat cases where a pattern does not match anything (or if there
are no test suites defined) as a hard failure, and exit unsuccessfully.