Mount data volumes if any are specified
Previously, if you tried to specify a data volume, it wouldn't mount
properly since we were using `HostConfig.Binds`. It turns out `Volumes`
is what you need (the online documentation is wrong or there is a bug;
I've submitted a pull request to fix the documentation here:
https://github.com/docker/docker/pull/20990)

This allows us to run Docker-in-Docker, since now you can simply specify
that a data volume should be created for `/var/lib/docker`. This will
create a volume on the host system, which won't be OverlayFS, and thus
you'll be able to use OverlayFS for the Docker daemon inside the
container.

At the end of the run, the container's data volumes will automatically
be removed, which is convenient for cleanup purposes.