Tag test suite images with their names
It is useful when viewing the output of `docker images` to see which
image was used in a test suite. This allows you to independently
(outside of navo) play around with an image by running:

    docker run --rm navo:suite_name command

Much easier than working with image IDs directly. This is more useful if
different Dockerfiles are used for different suites, since if you only
have a single Dockerfile you're just creating a bunch of tags to the
same image.

We could perhaps enhance this in the future to support pushing the
images to an arbitrary repo so that images can be shared across multiple
machines, though that's a micro-optimization that probably isn't
necessary.